fix test_anchor_generator by warp kernel launch parameter

+++ paddle/fluid/operators/detection/anchor_generator_op.cu
@@ -111,15 +111,15 @@ class AnchorGeneratorOpCUDAKernel : public framework::OpKernel<T> {
     framework::TensorFromVector(stride, ctx.device_context(), &sd);
 
     hipLaunchKernelGGL((GenAnchors<T>), dim3(grid), dim3(block), 0, stream,
-        anchors->data<T>(), ar.data<T>(), aspect_ratios.size(), as.data<T>(),
-        anchor_sizes.size(), sd.data<T>(), stride.size(), height, width,
+        anchors->data<T>(), ar.data<T>(), (const int)(aspect_ratios.size()), as.data<T>(),
+        (const int)(anchor_sizes.size()), sd.data<T>(), (const int)(stride.size()), (const int)(height), (const int)(width),
         offset);
 
     framework::Tensor v;
     framework::TensorFromVector(variances, ctx.device_context(), &v);
     grid = (box_num * 4 + block - 1) / block;
     hipLaunchKernelGGL((SetVariance<T>), dim3(grid), dim3(block), 0, stream, 
-         vars->data<T>(), v.data<T>(), variances.size(), box_num * 4);
+         vars->data<T>(), v.data<T>(), (const int)(variances.size()), (const int)(box_num * 4));
   }
 };  // namespace operators
